+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.6.2
+
+- Support specifying default slot in a Component with slots (e.g. `default_slot :some_slot`) to allow consumers to insert content into default slot by adding content directly under consumed component instead of opening a block with slot name explicitly
+- Support `:markup_root_slot` as a default slot that is always available to insert content into the component markup root element in case another element is specified as the default slot.
+- Hello, Component Listeners (Default Slot)! Sample (has simpler code than Hello, Component Listeners! due to specifying a default slot): `require 'glimmer-dsl-web/samples/hello/hello_component_listeners_default_slot'`
+
 ## 0.6.1
 
 - Hello, Component Listeners! Sample: `require 'glimmer-dsl-web/samples/hello/hello_component_listeners'`